+++ src/app/dashboard/page.tsx
@@ -6,7 +6,7 @@ import { RecipeSuggestions } from "@/components/dashboard/recipe-suggestions"
 import { categoriasDiferentes, quantidadeTotal } from "@/data/data"
 
 import { getServerSession } from "next-auth";
-import { authOptions } from "../api/auth/[...nextauth]/route";
+import { authOptions } from "@/lib/auth";
 import { redirect } from "next/navigation";
 import LogoutButton from "@/components/logoutButton"
